requirements.txt
pycryptodome
cryptography
Faker
rsa